refactor: Brand the web application
This implies change everything everywhere from mindgames to knowlift.
It's been a pleasure mindgames, what a journey it has been. Welcome knowlift.

Resolves #87
