+++ bin/webapp.py
@@ -30,8 +30,8 @@ import logging
 import web
 
 # Project specific
-from mindgames import settings
-from mindgames import urls
+from knowlift import settings
+from knowlift import urls
 
 logging.basicConfig(level=logging.INFO)
 logger = logging.getLogger(__name__)
@@ -41,7 +41,7 @@ app = web.application(urls.URLS, globals())
 
 def _load_session():
     """Initialize a session object using a particular type of storage (DiskStore, DBStore, etc.)"""
-    db = web.database(dbn='sqlite', db='mindgames.db')
+    db = web.database(dbn='sqlite', db='knowlift.db')
     store = web.session.DBStore(db, 'sessions')
     initializer = {'correct_answers': 0, 'incorrect_answers': 0, 'total_answers': 0, 'average': 5}
     session = web.session.Session(app, store, initializer)
