+++ docs/CONTRIBUTING.md
@@ -1,23 +1,23 @@
 ## Issues and Pull requests
-- If you want to file an issue please [perform a search](https://github.com/mariusmucenicu/mind-games/issues) to check whether anything similar already exists.
+- If you want to file an issue please [perform a search](https://github.com/mariusmucenicu/knowlift/issues) to check whether anything similar already exists.
   - If it does, add a comment to the existing issue instead of opening a new one.
-  - If there are multiple related issues, please reference them in the one you're commenting so that we can close the duplicates and keep the [issue board](https://github.com/mariusmucenicu/mind-games/issues) as organized and relevant as possible.
+  - If there are multiple related issues, please reference them in the one you're commenting so that we can close the duplicates and keep the [issue board](https://github.com/mariusmucenicu/knowlift/issues) as organized and relevant as possible.
 - Please don't file an issue to ask a question. For any questions feel free to <a href="mailto:marius_mucenicu@yahoo.com">contact the author</a>.
 
 ### Bug reports
-When [reporting a bug](https://github.com/mariusmucenicu/mind-games/issues) please include:
+When [reporting a bug](https://github.com/mariusmucenicu/knowlift/issues) please include:
 - Your device name and version, e.g `Samsung Galaxy S9, G960F` or `MacBook Pro (Retina, Mid 2015)`.
 - Your operating system name and version, e.g `macOS Mojave, Version 10.14.1`.
 - Your client name and version, e.g `Google Chrome, Version 70.0.3538.77`.
 - Detailed steps to reproduce the bug and any other details that might be helpful in troubleshooting.
 
 ### Feature requests
-When [proposing a feature](https://github.com/mariusmucenicu/mind-games/issues):
+When [proposing a feature](https://github.com/mariusmucenicu/knowlift/issues):
 - Explain in detail how it would work.
 - Keep the scope as narrow as possible, to make it easier to implement.
 
 ### Pull requests
-- Before submitting a pull request, please make sure you read the project's [Code style guidelines](https://github.com/mariusmucenicu/mind-games/blob/master/docs/CONTRIBUTING.md#code-style-guidelines) below.
+- Before submitting a pull request, please make sure you read the project's [Code style guidelines](https://github.com/mariusmucenicu/knowlift/blob/master/docs/CONTRIBUTING.md#code-style-guidelines) below.
 - After submitting a pull request, verify that all [status checks](https://help.github.com/articles/about-status-checks/) are passing.
 - After submitting a pull request, the reviewer(s) may ask you to complete additional work before your pull request can be ultimately accepted.
