+++ knowlift/lexicon.py
