+++ templates/layout.html
@@ -27,7 +27,7 @@ $def with (content)
         <div class="col-12">
           <div class="row hamburger-menu d-md-none">
             <div class="col-9 text-left">
-              <a class="font-weight-bold text-dark" href="/">MINDGAMES</a>
+              <a class="font-weight-bold text-dark" href="/">KNOWLIFT</a>
             </div>
             <div class="col-3 text-right">
               <i class="fas fa-bars fa-2x" onclick="openDrawer();"></i>
@@ -47,7 +47,7 @@ $def with (content)
               </div>
             </div>
             <div class="col-12 col-md-4 drawer-item">
-              <a href="/" class="font-weight-bold text-dark d-block d-md-inline">Mindgames</a>
+              <a href="/" class="font-weight-bold text-dark d-block d-md-inline">Knowlift</a>
             </div>
             <div class="col-12 col-md-4">
               <div class="row">
@@ -71,7 +71,7 @@ $def with (content)
 
       <div class="row footer flex-shrink-0">
         <div class="col-12 col-md-4 mb-2 border-top border-silver text-center">
-          &copy; 2018-2019 Mindgames <br>
+          &copy; 2018-2019 Knowlift <br>
           A pit stop for your brain.
         </div>
         <div class="col-12 col-md-4 mb-2 border-top border-silver text-center">
