+++ templates/legal.html
@@ -123,8 +123,8 @@ $var title: Legal
             </a>
           </div>
           <h2 class="fluid-subtitle">1. Terms and Conditions</h2>
-          <p class="text-justify">Please note that by accessing and using <strong>Mindgames</strong>
-            ("<strong>Mindgames</strong>", "<strong>our</strong>", "<strong>we</strong>",
+          <p class="text-justify">Please note that by accessing and using <strong>Knowlift</strong>
+            ("<strong>Knowlift</strong>", "<strong>our</strong>", "<strong>we</strong>",
             "<strong>us</strong>", "<strong>web application</strong>"), you consent unconditionally
             to these <strong>Terms and Conditions</strong> and to our
             <strong>Privacy Policy</strong> and <strong>Cookie Policy</strong>.</p>
@@ -138,20 +138,20 @@ $var title: Legal
             application can be improved, please
             <a href="mailto:marius_mucenicu@yahoo.com"><strong>contact the author</strong></a>.</p>
           <h3 class="fluid-subtitle" id="legal__liability">1.1 Liability</h3>
-          <p class="text-justify"><strong>Mindgames</strong> is not responsible for any injury,
+          <p class="text-justify"><strong>Knowlift</strong> is not responsible for any injury,
             loss, claim, or any direct, indirect, incidental, punitive, special, or consequential
             damages of any kind, including but without limitation to: loss of profits, loss of
             revenue, loss of savings, loss of data or any similar damages arising from your use of
             this web application or any information found on it.</p>
           <p class="text-justify mb-4">Every effort is made to keep the web application up and
-            running smoothly. However, <strong>Mindgames</strong>, takes no responsibility for, and
+            running smoothly. However, <strong>Knowlift</strong>, takes no responsibility for, and
             will not be liable for the web application being temporarily unavailable due to
             technical issues beyond our control.</p>
           <h3 class="fluid-subtitle" id="legal__litigation">1.2 Litigation</h3>
           <p class="text-justify">For any conflict that might arise between
-            <strong>Mindgames</strong> and its users we will seek to resolve promptly and amicably.
+            <strong>Knowlift</strong> and its users we will seek to resolve promptly and amicably.
             If, however, this is not possible, the litigation will be settled by a court of law in
-            the jurisdiction of <strong>Mindgames</strong>.</p>
+            the jurisdiction of <strong>Knowlift</strong>.</p>
         </div>
       </section>
 
@@ -170,7 +170,7 @@ $var title: Legal
           <h2 class="fluid-subtitle">2. Privacy Policy</h2>
           <p class="text-justify">This Privacy Policy covers the information we collect, store and
             use when you use our web application.</p>
-          <p class="text-justify"><strong>Mindgames</strong> collects the minimum amount of
+          <p class="text-justify"><strong>Knowlift</strong> collects the minimum amount of
             information necessary in order to provide you with the full application experience.
             Besides this, we also receive some standard information about the environment of a user,
             information which is part of the
@@ -355,16 +355,16 @@ $var title: Legal
             </a>
           </div>
           <h2 class="fluid-subtitle">4. Additional Information</h2>
-          <p class="text-justify"><u>Date of last revision: 2019.02.16</u></p>
+          <p class="text-justify"><u>Date of last revision: 2019.03.12</u></p>
           <p class="text-justify mb-4">Through this web application you are able to link to other
             websites/web applications, which are not under the control of
-            <strong>Mindgames</strong>. If you click on a third party link, you will be directed to
+            <strong>Knowlift</strong>. If you click on a third party link, you will be directed to
             that third party's site. We strongly advise you to review the
             <strong>Terms and Conditions</strong> and <strong>Privacy Policy</strong> of every site
             you visit. We have no control over the nature, content, accuracy of information,
             practices and availability of those sites/web applications.</p>
           <h3 class="fluid-subtitle" id="legal__history_of_software">4.1 History of software</h3>
-          <p class="text-justify mb-4"><strong>Mindgames</strong> was started in 2018 by
+          <p class="text-justify mb-4"><strong>Knowlift</strong> was started in 2018 by
             <a class="font-weight-bold"
                href="https://www.linkedin.com/in/mariusmucenicu/">Marius Mucenicu</a>, a graduate of
             the
