+++ tests/__init__.py
@@ -1,10 +1,10 @@
 """
-Main package for all mindgames tests.
+Main package for all knowlift tests.
 
 Modules:
 ========
-    lexicon_tests: Test mindgames.lexicon functionality.
-    number_distance_tests: Test mindgames.number_distance functionality.
+    lexicon_tests: Test knowlift.lexicon functionality.
+    number_distance_tests: Test knowlift.number_distance functionality.
     web_tests: Test bin.webapp functionality.
 
 Miscellaneous objects:
